# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateBrokerRequest:
    properties:
      clientId:
        type: string
      icon:
        $ref: '#/definitions/dto.Icon'
      isSsl:
        type: boolean
      keepAlive:
        type: integer
      name:
        type: string
      password:
        type: string
      port:
        type: integer
      server:
        type: string
      username:
        type: string
    required:
    - icon
    - isSsl
    - keepAlive
    - name
    - port
    - server
    type: object
  dto.CreateBrokerResponse:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
  dto.GetBrokerResponse:
    properties:
      clientId:
        type: string
      createdAt:
        type: string
      icon:
        $ref: '#/definitions/dto.Icon'
      id:
        format: uuid
        type: string
      isSsl:
        type: boolean
      keepAlive:
        type: integer
      name:
        type: string
      port:
        type: integer
      server:
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.Icon:
    properties:
      backgroundColor:
        type: string
      name:
        type: string
    required:
    - backgroundColor
    - name
    type: object
  dto.IconOptional:
    properties:
      backgroundColor:
        type: string
      name:
        type: string
    type: object
...
  dto.UpdateBrokerRequest:
    properties:
      clientId:
        type: string
      icon:
        $ref: '#/definitions/dto.IconOptional'
      isSsl:
        type: boolean
      keepAlive:
        type: integer
      name:
        type: string
      password:
        type: string
      port:
        type: integer
      server:
        type: string
      username:
        type: string
    type: object
...
  errors.HTTPError:
    properties:
      message:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    name: Deve-Lite
  title: DashboardX API
  version: "1.0"
paths:
  /brokers:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.GetBrokerResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.HTTPError'
      security:
      - BearerAuth: []
      summary: List brokers
      tags:
      - Brokers
    post:
      consumes:
      - application/json
      parameters:
      - description: Create data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBrokerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateBrokerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.HTTPError'
      security:
      - BearerAuth: []
      summary: Create a broker
      tags:
      - Brokers
  /brokers/{brokerId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Broker UUID
        in: path
        name: brokerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.HTTPError'
      security:
      - BearerAuth: []
      summary: Delete a broker
      tags:
      - Brokers
    get:
      consumes:
      - application/json
      parameters:
      - description: Broker UUID
        in: path
        name: brokerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetBrokerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.HTTPError'
      security:
      - BearerAuth: []
      summary: Get a broker
      tags:
      - Brokers
    patch:
      consumes:
      - application/json
      parameters:
      - description: Broker UUID
        in: path
        name: brokerId
        required: true
        type: string
      - description: Update data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateBrokerRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.HTTPError'
      security:
      - BearerAuth: []
      summary: Update a broker
      tags:
      - Brokers
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"